+++ packages/base/src/components/Buttons/IconButton/IconButton.tsx
@@ -1,0 +1,19 @@
+import type { ComponentPropsWithRef, ElementType } from 'react';
+import { forwardRef } from 'react';
+import { StyledIconButton } from './IconButton.styled';
+import type { IconButtonProps, IconButtonComponent } from './IconButton.types';
+
+const IconButton: IconButtonComponent = forwardRef(
+  <T extends ElementType = 'button'>(
+    { styleVar, sizeVar, disabled, as, children, ...rest }: IconButtonProps<T>,
+    ref: ComponentPropsWithRef<T>['ref'],
+  ) => {
+    return (
+      <StyledIconButton styleVar={styleVar} sizeVar={sizeVar} disabled={disabled} as={as} ref={ref} {...rest}>
+        {children}
+      </StyledIconButton>
+    );
+  },
+);
+
+export default IconButton;