+++ packages/base/src/components/Callout/index.ts
@@ -1,0 +1,3 @@
+export { default as Callout } from './Callout';
+export * from './Callout';
+export * from './Callout.types';