+++ packages/base/src/components/ControlButtons/MinusButton/index.ts
@@ -1,0 +1,2 @@
+export { default as MinusButton } from './MinusButton';
+export * from './MinusButton.types';