+++ packages/base/src/components/MinusButton/index.ts
@@ -1,3 +1,0 @@
-export { default as MinusButton } from './MinusButton';
-export * from './MinusButton.types';
-export * from './MinusButton.styled';