+++ packages/base/src/components/index.ts
@@ -3,4 +3,7 @@ export * from './Stack';
 export * from './Text';
 export * from './Modal';
 export * from './BackDrop';
+export * from './Switch';
 export * from './Chips';
+export * from './Buttons';
+export * from './ControlButtons';