+++ packages/base/src/styles/index.ts
@@ -8,8 +8,6 @@ import {
   typographyTokens,
 } from './tokens';
 
-// import './global.css';
-
 type ColorTokens = keyof typeof colorTokens;
 type SpacingTokens = keyof typeof spacingTokens;
 type BorderRadiusTokens = keyof typeof borderRadiusTokens;