+++ test-env/src/App.tsx
@@ -3,22 +3,34 @@ import './App.css';
 
 import reactLogo from './assets/react.svg';
 import viteLogo from '/vite.svg';
-import styled from '@emotion/styled';
-import { Modal, useHandleModal, ChipToggle } from '@shoplflow/base';
+import { Modal, useHandleModal, ChipToggle, Stack, Text } from '@shoplflow/base';
 
 const TestModal = () => {
-  const { removeModal } = useHandleModal();
+  const { removeModal, addModal } = useHandleModal();
   const array = new Array(200).fill(0);
   return (
-    <Modal.Container
-      outsideClick={() =>
-        removeModal({
-          deps: 1,
-        })
-      }
-    >
+    <Modal.Container>
       <Modal.Header>Header</Modal.Header>
       <Modal.Body>
+        <button
+          onClick={() =>
+            removeModal({
+              deps: 0,
+            })
+          }
+        >
+          remove deps
+        </button>
+        <button
+          onClick={() =>
+            removeModal({
+              id: 'TestModal',
+            })
+          }
+        >
+          remove id
+        </button>
+        <button onClick={() => addModal(<TestModal />)}>add</button>
         {array.map((_, index) => (
           <ChipToggle text={'Toggle' + index} key={index} />
         ))}
@@ -31,35 +43,30 @@ const TestModal = () => {
 function App() {
   const [count, setCount] = useState(0);
 
-  const SomeText = styled.div`
-    color: var(--primary400);
-    box-shadow: var(--dropShadow);
-  `;
-
   const { addModal } = useHandleModal();
 
   return (
-    <>
-      <div>
+    <Stack.Vertical align={'center'}>
+      <Stack.Horizontal width={'100%'} justify={'center'}>
         <a href='https://vitejs.dev' target='_blank' rel='noreferrer noopener'>
           <img src={viteLogo} className='logo' alt='Vite logo' />
         </a>
         <a href='https://react.dev' target='_blank' rel='noreferrer noopener'>
           <img src={reactLogo} className='logo react' alt='React logo' />
         </a>
-      </div>
-      <SomeText>Some Text</SomeText>
-      <h1 className={'heading1_700'}>Vite + React</h1>
-
-      <div className='card'>
-        <button onClick={() => setCount((count) => count + 1)}>count is {count}</button>
-        <button onClick={() => addModal(<TestModal />)}>addModal</button>
+      </Stack.Horizontal>
+      <h1 className={'heading1_700'}>Shoplflow Test Env</h1>
+      <Text>some Test</Text>
+      <Stack.Vertical align={'center'}>
+        <Stack.Horizontal justify={'center'}>
+          <button onClick={() => setCount((count) => count + 1)}>count is {count}</button>
+          <button onClick={() => addModal(<TestModal />, 'TestModal')}>addModal</button>
+        </Stack.Horizontal>
         <p>
           Edit <code>src/App.tsx</code> and save to test HMR
         </p>
-      </div>
-      <p className='read-the-docs'>Click on the Vite and React logos to learn more</p>
-    </>
+      </Stack.Vertical>
+    </Stack.Vertical>
   );
 }